Refactor to move follow-up summaries into lib
